python-services/ocr-service/requirements.txt
# # # flask==2.3.3
# # # flask-cors==4.0.0



# # flask==2.3.3
# # flask-cors==4.0.0
# # pytesseract==0.3.10
# # pillow==10.0.1
# # opencv-python==4.8.1.78
# # requests==2.31.0
# # pytesseract==0.3.10


# flask==2.3.3
# flask-cors==4.0.0
# easyocr==1.7.0
# pillow==10.0.1
# opencv-python==4.8.1.78
# requests==2.31.0
# numpy==1.24.3

flask==2.3.3
flask-cors==4.0.0
easyocr==1.7.0
pillow==10.0.1
opencv-python==4.8.1.78
requests==2.31.0
numpy==1.24.3
torch==2.0.1
torchvision==0.15.2